:recycle: util code 정리
util code 정리
